# openapi copy.yaml
openapi: 3.0.0
info:
  title: 图书馆智能AI系统 API
  description: 图书馆智能AI系统的API接口文档
  version: 1.0.0
servers:
- url: http://172.20.0.3:8888
  description: 本地开发服务器
paths:
  /lib/can_book_seat:
    get:
      summary: 获取用户能否预约（含全媒体）,通过用户学工号获取是否有预约（含全媒体）权限
      description: ''
      operationId: can_book_seat
      parameters:
      - required: true
        in: query
        description: 用户学工号
        name: user_id
        type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/definitions/ReserveCheckResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
  /lib/check_book_can_renew:
    post:
      summary: 检查图书是否可以续借,根据条码判断图书是否可以被续借
      description: ''
      operationId: check_book_can_renew
      requestBody:
        required: true
        description: 包含book_id数组的JSON对象
        content:
          application/json:
            schema:
              type: object
              properties:
                book_id:
                  type: array
                  items:
                    type: integer
                  example: [32086180]
                  description: 书籍条码列表
                  not:
                    type: string
              required:
                - book_id
      # parameters:
      # - required: true
      #   in: query
      #   description: 图书条码
      #   name: book_id
      #   type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/definitions/RenewCheckResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
  /lib/get_user_borrow_info:
    get:
      summary: 获取用户借阅信息,通过用户学工号获取在借所有资源的信息，根据应还时间来判断是否过期/逾期
      description: ''
      operationId: get_user_borrow_info
      parameters:
      - required: true
        in: query
        description: 用户学工号
        name: user_id
        type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/definitions/BorrowInfoResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
  /lib/get_user_info:
    get:
      summary: 获取用户信息接口,通过用户学工号查询用户信息，包含姓名、学工号、系统ID、物理卡号、NFC ID、邮箱、电话、进馆权限、借阅权限、部门、身份有效期、通讯有效期、身份类型、逾期费等信息
      description: ''
      operationId: get_user_info
      parameters:
      - required: true
        in: query
        description: 用户学工号
        name: user_id
        type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: {}
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  success:
                    type: boolean
components:
  schemas:
    Error:
      properties:
        message:
          type: string
          description: 错误信息
        success:
          type: boolean
          description: 错误状态
      type: object
    BorrowInfoResponse:
      properties:
        data:
          $ref: '#/definitions/BorrowInfoList'
        message:
          type: string
          description: 响应信息
        success:
          type: boolean
          description: 响应状态
      type: object
    BorrowInfoList:
      properties:
        borrow_info:
          type: array
          items:
            $ref: '#/definitions/BorrowInfo'
      type: object
    BorrowInfo:
      properties:
        source:
          type: string
          description: 数据来源
        barcode:
          type: string
          description: 条码
        book_name:
          type: string
          description: 题名
        book_author:
          type: string
          description: 责任者
        publish:
          type: string
          description: 出版社
        pub_year:
          type: string
          description: 出版年份
        sub_library_code:
          type: string
          description: 分馆代码
        collection_code:
          type: string
          description: 馆藏地代码
        borrow_time:
          type: string
          description: 借阅时间
        should_return_time:
          type: string
          description: 应还时间
        call_no:
          type: string
          description: 索书号
        isbn:
          type: string
          description: isbn号
        material:
          type: string
          description: 介质类型
        overdue:
          type: boolean
          description: 是否逾期
      type: object
    ReserveCheckResponse:
      properties:
        data:
          $ref: '#/definitions/ReserveCheck'
        message:
          type: string
          description: 响应信息
        success:
          type: boolean
          description: 响应状态
      type: object
    ReserveCheck:
      properties:
        can_book_seat:
          type: boolean
          description: 是否可以预约
      type: object
    RenewCheckResponse:
      properties:
        data:
          $ref: '#/definitions/RenewCheck'
        message:
          type: string
          description: 响应信息
        success:
          type: boolean
          description: 响应状态
      type: object
    RenewCheck:
      properties:
        can_renew:
          type: boolean
          description: 是否可以续借
        call_no:
          type: string
          description: 索书号
        collection_cod:
          type: string
          description: 馆藏地代码
        renew_description:
          type: string
          description: 续借说明
        renew_status:
          type: string
          description: 是否可以续借状态
        sublibrary:
          type: string
          description: 分馆代码
        title:
          type: string
          description: 题名
      type: object